 requirements.txt
streamlit
faiss-cpu
langchain
PyPDF2
llama-cpp-python
python-dotenv
tiktoken
SpeechRecognition
